Move new modules into a new modules directory and rename some contracts
