+++ contracts/modules/DecentHatsCreationModule.sol
@@ -5,16 +5,16 @@ import {Enum} from "@gnosis.pm/safe-contracts/contracts/common/Enum.sol";
 import {IAvatar} from "@gnosis.pm/zodiac/contracts/interfaces/IAvatar.sol";
 import {Strings} from "@openzeppelin/contracts/utils/Strings.sol";
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import {IERC6551Registry} from "./interfaces/IERC6551Registry.sol";
-import {IHats} from "./interfaces/hats/IHats.sol";
-import {LockupLinear, Broker} from "./interfaces/sablier/types/DataTypes.sol";
-import {IHatsModuleFactory} from "./interfaces/hats/IHatsModuleFactory.sol";
-import {IHatsElectionsEligibility} from "./interfaces/hats/modules/IHatsElectionsEligibility.sol";
+import {IERC6551Registry} from "../interfaces/IERC6551Registry.sol";
+import {IHats} from "../interfaces/hats/IHats.sol";
+import {LockupLinear, Broker} from "../interfaces/sablier/types/DataTypes.sol";
+import {IHatsModuleFactory} from "../interfaces/hats/IHatsModuleFactory.sol";
+import {IHatsElectionsEligibility} from "../interfaces/hats/modules/IHatsElectionsEligibility.sol";
 import {ModuleProxyFactory} from "@gnosis.pm/zodiac/contracts/factory/ModuleProxyFactory.sol";
-import {ISablierV2LockupLinear} from "./interfaces/sablier/ISablierV2LockupLinear.sol";
-import {DecentHatsUtils} from "./DecentHatsUtils.sol";
+import {ISablierV2LockupLinear} from "../interfaces/sablier/ISablierV2LockupLinear.sol";
+import {DecentHatsModuleUtils} from "./DecentHatsModuleUtils.sol";
 
-contract DecentHatsCreationModule is DecentHatsUtils {
+contract DecentHatsCreationModule is DecentHatsModuleUtils {
     struct TopHatParams {
         string details;
         string imageURI;
