+++ contracts/modules/DecentSablierStreamManagementModule.sol
@@ -3,12 +3,10 @@ pragma solidity 0.8.28;
 
 import {Enum} from "@gnosis.pm/safe-contracts/contracts/common/Enum.sol";
 import {IAvatar} from "@gnosis.pm/zodiac/contracts/interfaces/IAvatar.sol";
-import {ISablierV2Lockup} from "./interfaces/sablier/ISablierV2Lockup.sol";
-import {Lockup} from "./interfaces/sablier/types/DataTypes.sol";
+import {ISablierV2Lockup} from "../interfaces/sablier/ISablierV2Lockup.sol";
+import {Lockup} from "../interfaces/sablier/types/DataTypes.sol";
 
-contract DecentSablierStreamManagement {
-    string public constant NAME = "DecentSablierStreamManagement";
-
+contract DecentSablierStreamManagementModule {
     function withdrawMaxFromStream(
         ISablierV2Lockup sablier,
         address recipientHatAccount,
