+++ deploy/core/018_deploy_DecentSablierStreamManagement.ts
@@ -3,7 +3,7 @@ import { DeployFunction } from 'hardhat-deploy/types';
 import { deployNonUpgradeable } from '../helpers/deployNonUpgradeable';
 
 const func: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
-  await deployNonUpgradeable(hre, 'DecentSablierStreamManagement');
+  await deployNonUpgradeable(hre, 'DecentSablierStreamManagementModule');
 };
 
 export default func;
